pyproject.toml
[project]
name = "notification-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django>=5.2.5",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "django-sms>=0.7.0",
    "djangorestframework>=3.16.1",
    "environs>=14.3.0",
    "factory-boy>=3.3.3",
    "pytelegrambotapi>=4.28.0",
    "pytest>=8.4.1",
    "pytest-mock>=3.14.1",
    "requests>=2.32.5",
]
